refactor: split executor upstream and persistence helpers
