+++ crates/agentic-core/src/executor/mod.rs
@@ -10,6 +10,7 @@ pub mod rehydrate;
 pub mod request;
 
 mod gateway;
+mod upstream;
 
 pub use engine::{BoxStream, ExecuteRequest, create_conversation, execute};
 pub use error::{ExecutorError, ExecutorResult};
